Crate kidou

This binary crate is a command-line utility to integrate the orbits of globular clusters by providing their initial positions and velocities and choosing a model of the Galaxy potential. It is mostly based on the paper of Bajkova & Bobylev (2020, v1).

Module documentations:

Command-line interface

Description of units conversions before and after the integration

A list of available potentials and their derivatives

A list of available models

Modules

cli

This module provides the CLI routines

orbit

This module defines the Orbit struct

Constants

PADDING

Padding used when printing the output

PI

The number $ \pi $ casted to the floating point type

Functions

main

Run the program

Type Definitions

F

The floating point type used across the program

I

The integer type used across the program