Trait kidou::orbit::calc::integrate::constants::correction::Corrections

pub trait Corrections {
    type Output;
    fn to_gca_x(&self) -> Self::Output;

    fn to_gca_z(&self) -> Self::Output;

    fn to_gca_u(&self) -> Self::Output;

    fn to_gca_v(&self) -> Self::Output;

    fn to_gca_w(&self) -> Self::Output;
}

Adds correction methods to float types

Associated Types

type Output

A float type. Should be the same as Self.

Required methods

fn to_gca_x(&self) -> Self::Output

Convert X component of the radius vector from the Heliocentric Cartesian system to Galactic Cylindrical system

fn to_gca_z(&self) -> Self::Output

Convert Z component of the radius vector from the Heliocentric Cartesian system to Galactic Cylindrical system

fn to_gca_u(&self) -> Self::Output

Convert U component of the velocity vector from the Heliocentric Cartesian system to Galactic Cylindrical system

fn to_gca_v(&self) -> Self::Output

Convert V component of the velocity vector from the Heliocentric Cartesian system to Galactic Cylindrical system

fn to_gca_w(&self) -> Self::Output

Convert W component of the velocity vector from the Heliocentric Cartesian system to Galactic Cylindrical system