Module kidou::orbit::calc::models

This module provides general models of the Galactic potential

There are two groups of models provided, depending on how they represent the Galactic potential:

A sum of potentials: bulge + disk + halo;

A sum of potentials: bulge + thin disk + thick disk + halo.

These are combinations of the available potentials.

Group 1

#	Bulge	Disk	Halo	Values of the parameters

1	P	MN	NFW	Bajkova, Bobylev (2020, v1)

Group 2

#	Bulge	Thin Disk	Thick Disk	Halo	Values of the parameters

2	P	MN	MN	NFW	Pouliasis et al. (2017, model I) + Eilers (2018)

Modules

m1

This module provides the M1 model

m2

This module provides the M2 model

model

This module provides the Model trait

Structs

M1

This model uses the Plummer potential for bulge, the Miyamoto & Nagai potential for disk, and the Navarro-Frenk-White potential for halo. The values of the parameters are taken from Bajkova, Bobylev (2020, v1).

M2

This model uses the Plummer potential for bulge, the Miyamoto & Nagai potential for thin and thick disks, and the Navarro-Frenk-White potential for halo. The values of the parameters for the bulge, the thin disk and the thick disk are taken from Pouliasis et al. (2017, model I), values of the parameters for the halo are taken from Eilers (2018).

Constants

MODELS

An array of models: indices of the items are the same as the indices of the models

Traits

Model

This trait defines that a model needs to provide functions that return the values of Galactic potential and its derivatives.